+++ generators/activity/index.js
@@ -50,4 +50,3 @@ module.exports = generators.Base.extend( {
    }
 
 } );
-