+++ generators/app/templates/demo.includes/widgets/example/my-navigation-widget/spec/my-navigation-widget.spec.js
@@ -7,7 +7,7 @@ define( [
    'use strict';
 
    // More information on widget tests:
-   // https://github.com/LaxarJS/laxar-mocks/blob/master/docs/manuals/index.md
+   // https://laxarjs.org/docs/laxar-mocks-v1-latest/manuals/
    describe( 'The MyNavigationWidget', function() {
 
       beforeEach( axMocks.createSetupForWidget( descriptor, {