+++ generators/widget/index.js
@@ -214,7 +214,7 @@ module.exports = generators.Base.extend( {
    end: function() {
       'use strict';
       if( this.placeholder.laxarIntegrationTechnology === 'react' ) {
-         var link = 'https://github.com/LaxarJS/laxar-react-adapter/blob/master/README.md';
+         var link = 'https://laxarjs.org/docs/laxar-react-adapter-v0-latest/';
          this.log( '\nEnsure that your application is prepared for using the ' +
                    this.placeholder.laxarIntegrationTechnology + ' adapter or do the necessary steps which' +
                    ' are described in the documentation of the adapter:' );
@@ -225,6 +225,6 @@ module.exports = generators.Base.extend( {
          this.log( '\nYou can now start developing your ' + this.placeholder.artifactType + '!' );
       }
       this.log( 'For more information about developing widgets with LaxarJS, please refer to the manuals:' );
-      this.log( '\nhttps://github.com/LaxarJS/laxar/blob/master/docs/manuals/index.md#manuals' );
+      this.log( '\nhttps://laxarjs.org/docs/laxar-v1-latest/manuals/' );
    }
 } );