+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Last Changes
 
+- [#21](https://github.com/LaxarJS/generator-laxarjs/issues/21): generators: changed links of documentation from github to new documentation site
 - [#25](https://github.com/LaxarJS/generator-laxarjs/issues/25): docs: Add hint about arguments when using a local installation of yo
 - [#23](https://github.com/LaxarJS/generator-laxarjs/issues/23): generators: updated version of dependency in bower.json
 - [#19](https://github.com/LaxarJS/generator-laxarjs/issues/19): generators: fixed license validation and filter for inquirer v1.0.0